+++ classgraph/src/main/java/io/github/classgraph/Scanner.java
@@ -462,7 +462,7 @@ class Scanner implements Callable<ScanResult> {
         // to strip the "!/" off the end. Also strip any "jar:file:" or "file:" off the beginning. This normalizes
         // "file:x.jar" and "x.jar" to the same string, for example.
         if (!(classpathEntryObjNormalized instanceof Path)) {
-            classpathEntryObjNormalized = FastPathResolver.resolve(FileUtils.currDirPath(),
+            classpathEntryObjNormalized = FastPathResolver.resolveFilePath(FileUtils.currDirPath(),
                     classpathEntryObjNormalized.toString());
         }